+++ packages/files/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.13](https://github.com/standardnotes/app/compare/@standardnotes/files@1.15.12...@standardnotes/files@1.15.13) (2023-06-27)
+
+**Note:** Version bump only for package @standardnotes/files
+
 ## [1.15.12](https://github.com/standardnotes/app/compare/@standardnotes/files@1.15.11...@standardnotes/files@1.15.12) (2023-06-25)
 
 **Note:** Version bump only for package @standardnotes/files