+++ packages/models/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.46.3](https://github.com/standardnotes/app/compare/@standardnotes/models@1.46.2...@standardnotes/models@1.46.3) (2023-06-27)
+
+**Note:** Version bump only for package @standardnotes/models
+
 ## [1.46.2](https://github.com/standardnotes/app/compare/@standardnotes/models@1.46.1...@standardnotes/models@1.46.2) (2023-06-25)
 
 **Note:** Version bump only for package @standardnotes/models