+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.332](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.331...@standardnotes/releases@1.4.332) (2023-06-27)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.331](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.330...@standardnotes/releases@1.4.331) (2023-06-26)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.330](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.329...@standardnotes/releases@1.4.330) (2023-06-26)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.329](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.328...@standardnotes/releases@1.4.329) (2023-06-26)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.4.328](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.327...@standardnotes/releases@1.4.328) (2023-06-25)
 
 **Note:** Version bump only for package @standardnotes/releases