+++ packages/snjs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.169.5](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.169.4...@standardnotes/snjs@2.169.5) (2023-06-27)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
+## [2.169.4](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.169.3...@standardnotes/snjs@2.169.4) (2023-06-26)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
 ## [2.169.3](https://github.com/standardnotes/app/compare/@standardnotes/snjs@2.169.2...@standardnotes/snjs@2.169.3) (2023-06-25)
 
 ### Bug Fixes